requirements.txt
# Bedrock & Strands Agents
strands-agents>=0.2.0
strands-agents-tools>=0.2.0

# OAuth2 & Authentication
authlib>=1.3.0
python-jose[cryptography]>=3.3.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AWS SDK & CDK
boto3>=1.28.0
botocore>=1.31.0
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# Data Handling & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0
python-json-logger>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
